+++ module/actions/roll-dialog.js
@@ -1,22 +1,25 @@
 import { HLMApplication } from "../sheets/application.js";
-import { ATTRIBUTES, COVER_STATES } from "../constants.js";
+import { ATTRIBUTES, COVER_STATES, ROLL_MODIFIER_TYPE, ATTRIBUTE_MIN, ATTRIBUTE_MAX_ROLLED } from "../constants.js";
 import { Utils } from "../utilities/utils.js";
 
 export class RollElement {
     value
     type
     description
+    classification
 
     /**
      * Represents a modifier to a die roll
      * @param {integer} value The value of the modifier
      * @param {string} type "die" or "flat"
      * @param {string} description The description to show to the player
+     * @param {string} classification "modifier" or "bonus"
      */
-    constructor(value, type, description) {
+    constructor(value, type, description, classification) {
         this.value=value;
         this.type=type;
         this.description=description;
+        this.classification=classification;
     }
 }
 
@@ -94,34 +97,48 @@ export class RollDialog extends HLMApplication {
         let formData=super._getSubmitData(data);
     }
 
-    triggerRoll() {
+    totalAttributes() {
         this.dieModifiers.push({
             value: this.additionalDie,
             type: "die",
-            description: "Additional"
+            description: "Additional",
+            classification: "bonus"
         });
         this.flatModifiers.push({
             value: this.additionalFlat,
-            type: "die",
-            description: "Additional"
+            type: "flat",
+            description: "Additional",
+            classification: "bonus"
         });
 
         let totalDieCount=0;
-        let totalFlat=0;
+        let totalAttr=0;
+        let totalBonus=0;
 
         this.dieModifiers.forEach((modifier) => {
             totalDieCount+=parseInt(modifier.value);
         });
         this.flatModifiers.forEach((modifier) => {
-            totalFlat+=parseInt(modifier.value);
+            if(modifier.classification == ROLL_MODIFIER_TYPE.modifier) {
+                totalAttr+=parseInt(modifier.value);
+            } else if(modifier.classification == ROLL_MODIFIER_TYPE.bonus) {
+                totalBonus+=parseInt(modifier.value);
+            }
         });
+        if(totalAttr<ATTRIBUTE_MIN) totalAttr=ATTRIBUTE_MIN;
+        if(totalAttr>ATTRIBUTE_MAX_ROLLED) totalAttr=ATTRIBUTE_MAX_ROLLED;
         if(this.focused) {
             totalDieCount+=1;
         }
+        return {"dice": totalDieCount, "attribute": totalAttr, "bonus": totalBonus}
+    }
+
+    async triggerRoll() {
+        const modifierTotals=this.totalAttributes();
         if(this.internal) {
-            this.actor.triggerRolledInternal(this.internal,this.attribute,totalDieCount,totalFlat,this.cover);
+            await this.actor.triggerRolledInternal(this.internal,this.attribute,modifierTotals.dice,modifierTotals.attribute+modifierTotals.bonus,this.cover);
         } else {
-            this.actor.rollAttribute(this.attribute,totalDieCount,totalFlat, this.cover);
+            await this.actor.rollAttribute(this.attribute,modifierTotals.dice,modifierTotals.attribute+modifierTotals.bonus, this.cover);
         }
         this.close();
     }